+++ pages/employees/add.vue
@@ -1,0 +1,4 @@
+<template>
+  <employee-form mode="add" />
+</template>
+
